+++ src/sw/allotmentclub/alembic/versions/add_assignment_todo_9da4a1cc4bf1.py
@@ -1,0 +1,30 @@
+"""Add assignment todo
+
+Revision ID: 9da4a1cc4bf1
+Revises: 48303c2cf2eb
+Create Date: 2023-07-30 13:32:35.422103
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+# revision identifiers, used by Alembic.
+revision = '9da4a1cc4bf1'
+down_revision = '48303c2cf2eb'
+
+
+def upgrade():
+    op.create_table(
+        'assignmenttodo',
+        sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('organization_id', sa.Integer(), nullable=False),
+        sa.Column('description', sa.String(), nullable=True),
+        sa.Column(
+            'priority', sa.Integer(), nullable=False, server_default='4'
+        ),
+        sa.PrimaryKeyConstraint('id')
+    )
+
+
+def downgrade():
+    op.drop_table('assignmenttodo')